+++ index.html
@@ -38,5 +38,5 @@
     </div>
   </body>
 
-  <script src="index.js"></script>
+  <script type="module" src="index.js"></script>
 </html>
